Change to State Management
This version is a significant change in state management, as well as the addition of logic for an OLED SH1106 screen.  Toggles for the OLED and Serial monitor printouts have been "jiggered" so that they can be enabled/disabled.  When disabled and compiled these can add significant memory savings for the arduino.  This version changes also set the groundwork for a branchless version in the future.  The original version is still entirely effective and far less verbose and so still retains value, thus, the second version is uploaded seperately.